02-similarity/codespace/pyproject.toml
[project]
name = "codespace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.20.2",
    "fastapi>=0.128.0",
    "matplotlib>=3.10.8",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.0",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.40.0",
]
